update price decimal

+++ contracts/interfaces/PriceOracleInterface.sol
@@ -6,7 +6,7 @@ interface PriceOracleInterface {
     event DecimalsUpdate(uint256 chainID, uint256 decimal);
     event InitCurrencyInfo(uint256 chainID, uint256 price, uint256 decimal);
 
-    function setPrice(uint256 chainID, uint112 price) external;
+    function setPrice(uint256 chainID, uint128 price) external;
 
     function getPrice(uint256 chainID) external view returns (uint256 price);
 
@@ -19,7 +19,7 @@ interface PriceOracleInterface {
 
     function initCurrencyInfo(
         uint256 chainID,
-        uint112 price,
+        uint128 price,
         uint64 decimal
     ) external;
